move array helpers to linking-to header
